requirements.txt
mailgun
requests
pandas
boto3
matplotlib
psycopg2-binary
notebook
langchain_aws
langchain_community
langgraph
awscli
python-dotenv
langgraph-cli[inmem]
ipykernel
#pygraphviz
sqlalchemy
uvicorn
fastapi
pydantic
python-multipart
psycopg
premailer
gunicorn
langchain

pydantic
rich
python-dotenv
fastapi
uvicorn
python-multipart
sqlalchemy
jinja2
boto3
pytest
pytest-asyncio
alembic
psycopg2-binary
requests
authlib
itsdangerous